refactor(galgame): split plugin_capture_profile_helpers.py from __init__.py
Move 9 capture profile helpers verbatim:
- _normalize_ocr_capture_profile_stage
- _normalize_ocr_capture_profile_save_scope
- _is_ratio_profile_payload
- _normalize_ocr_capture_profile_payload
- _capture_profile_entry_to_stage_map
- _capture_profile_bucket_entry_to_stage_map
- _capture_profile_entry_to_window_bucket_map
- _window_bucket_map_to_capture_profile_payload
- _capture_profile_components_to_entry

__init__.py re-exports them so external import paths stay compatible.
